Add `apt update` back in, it seems unlikely this is causing problems

+++ packer/ubuntu/update.sh
@@ -18,4 +18,4 @@
 set -x -e -o pipefail
 
 export DEBIAN_FRONTEND=noninteractive
-apt-get -y autoremove && apt remove -y unattended-upgrades
+apt-update && apt-get -y autoremove && apt remove -y unattended-upgrades
